Prototype Town (buildings not yet completed)
Marissa, I've changed the boundaries of the town (where each building
begins and ends) so that everything is nice and symmetrical. This should
make calculating everything easier, seeing as how the offset from each
side of the screen will be the same. However, some adjustments to your
code will need to be made.